+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0
 
 - Enable Node.js environment for `*.config.ts` files
+- Add `denyWarnings`, `reportUnusedDisableDirectives` and `typeAware` options to TypeScript config
 - Fix missing explicit OxlintConfig type annotations to exported configs ([#4](https://github.com/rivuty/oxlint-config/issues/4))
 
 ## 1.1.0